workaround conversion warning

+++ Surface_mesh_shortest_path/include/CGAL/Surface_mesh_shortest_path/Surface_mesh_shortest_path.h
@@ -1268,7 +1268,7 @@ private:
       propagateRight = rightSide;
     }
 
-    if (node->level() <= num_faces(m_graph))
+    if (node->level() <= static_cast<std::size_t>(num_faces(m_graph)))
     {
       if (propagateLeft)
       {
